Вынесла парсинг строки запроса в отдельную функцию
