+++ src/constants/sortOptions.ts
@@ -1,3 +1,3 @@
-import { SortItem } from "@/redux/filterSlice";
+import { SortItem } from "@/types/filterTypes";
 
 export const SORT_OPTIONS: SortItem[] = [{name: 'цене', sort: 'price'}, {name: 'популярности', sort: 'rating'}, {name: 'алфавиту', sort: 'title'}];
