+++ src/hooks/useSetQueryParams.ts
@@ -1,7 +1,6 @@
 import { useGetAllProductsQuery } from "@/api/product/productApi";
-import { FilterSliceState, PaginationType, SortItem } from "@/redux/filterSlice";
-// import { fetchProducts } from "@/redux/productsSlice";
 import { useAppDispatch } from "@/redux/store";
+import { FilterSliceState, PaginationType, SortItem } from "@/types/filterTypes";
 import qs from 'qs';
 import { RefObject, useEffect } from "react";
 import { useNavigate } from "react-router-dom";
@@ -10,12 +9,9 @@ type SelectedFilters = FilterSliceState['selected'];
 
 function useSetQueryParams( sortTypeValue: SortItem, categoryId: number, pagination: PaginationType, selected: SelectedFilters, isFirstMountRef: RefObject<boolean> ) {
 
-    const dispatch = useAppDispatch();
     const navigate = useNavigate();
 
     useEffect(() => {
-
-        // Параметры отображаются в строке запроса, когда поменяется одна из зависимостей
 
         if(!isFirstMountRef.current) {
             const query = qs.stringify({
@@ -48,20 +44,6 @@ function useSetQueryParams( sortTypeValue: SortItem, categoryId: number, paginat
 
     return { data, error, isLoading };
 
-    // useEffect(() => {
-    //     dispatch(
-    //         fetchProducts({
-    //             sortTypeValue, 
-    //             categoryId, 
-    //             pagination: {
-    //                 currentPage: pagination.currentPage,
-    //                 limit: pagination.limit
-    //             },
-    //             selected
-    //         })
-    //     );
-        
-    // }, [sortTypeValue, categoryId, pagination.currentPage, pagination.limit, selected]);
 }
 
 export default useSetQueryParams;
