+++ src/lib/normalizeToArray.ts
@@ -1,0 +1,5 @@
+export const normalizeToArray = (value: unknown): string[] => {
+    if (Array.isArray(value)) return value;
+    if (typeof value === 'string') return [value];
+    return [];
+};
